requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
wheel
psycopg2-binary
flask-cors
google-cloud-aiplatform
google-cloud-speech
google-cloud-texttospeech
google-cloud-vertex-ai
speechrecognition
pyttsx3
vertexai